+++ config/environments/staging.rb
@@ -90,4 +90,7 @@ Rails.application.configure do
 
   # FITS characterization
   config.run_fits_characterization = true
+
+  # Action on unpermitted parameters
+  config.action_controller.action_on_unpermitted_parameters = false
 end