+++ config/environments/uat.rb
@@ -91,4 +91,7 @@ Rails.application.configure do
 
   # FITS characterization
   config.run_fits_characterization = false
+
+  # Action on unpermitted parameters
+  config.action_controller.action_on_unpermitted_parameters = false
 end